backend/requirements.txt
web3>=6.0.0
crewai
python-dotenv
